refactor: share cursor argument parsing without Relay dependency
