+++ packages/core/src/utils/index.ts
@@ -21,6 +21,7 @@ import {
 } from '../types/index.js';
 
 export * from './base64.js';
+export * from './connection-args.js';
 export * from './context-cache.js';
 export * from './cursors.js';
 export * from './enums.js';
